Move task specific transition extenders to opengever.task.
